+++ sys/netns/spp_usrreq.c
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)spp_usrreq.c	8.1 (Berkeley) 06/10/93
+ *	@(#)spp_usrreq.c	8.2 (Berkeley) 01/09/95
  */
 
 #include <sys/param.h>
@@ -1241,7 +1241,7 @@ spp_usrreq(so, req, m, nam, controlp)
 	register struct sockbuf *sb;
 
 	if (req == PRU_CONTROL)
-                return (ns_control(so, (int)m, (caddr_t)nam,
+                return (ns_control(so, m, (caddr_t)nam,
 			(struct ifnet *)controlp));
 	if (nsp == NULL) {
 		if (req != PRU_ATTACH) {
